refactor: split Option family into src/results/option.py (#13)
Phase 1 structural cut: move Option/Some/Null out of results.py into a new
src/results/option.py. Result/Ok/Err and the error hierarchy stay in
results.py (no results.py->result.py rename). Public API is byte-identical:
__init__.py re-exports every name from both modules.

The results<->option import cycle is broken with function-local (deferred)
imports of Ok/Err/UnwrapFailedError inside the Option methods that construct
them; results.py keeps a top-level `from .option import Null, Some`. This
keeps every Result/Ok/Err method body byte-identical (disjoint with two
sibling lanes editing Ok.is_err_and / Err.unwrap) — results.py diff is limited
to top-of-file imports and deletion of the moved block.

The Some.__init__ None guard moves intact; test_public_api.py pins the public
import surface and re-confirms the boundary guard fires post-move.
